recreate address logic
