--- address-servicee/src/routes.js
@@ -1,7 +1,0 @@
-const { Router } = require("express");
-const AddressController = require("./controllers/address-controller");
-const routes = new Router();
-
-routes.post("/address", AddressController.search);
-
-module.exports = routes;
