--- address-servicee/src/services/address-service.js
@@ -1,7 +1,0 @@
-class AddressService {
-    static search(user) {
-        return None // Nao sei o que vai ser aqui
-    }
-}
-
-module.exports = AddressService;
